split MetadataTest into two distinct functionality descriptions:
* general metadata usage
* behavior of store attribute of MetaClass
